Really really sweeping changes to the Stegdienst app!
They're too numerous to list them all. Basically, the app is now fully usable for 2014. Import/Export is implemented and seems to work great, there has been further UI clean-up, more strategies have been implemented, a number of bugs fixed (particularly the nasty "ghost row" issue that prevented more than one suggestion in the same session), and most warnings are implemented. All in all we're ready for the public preview!

However, there are still a number of serious problems that prevent starting work on the Monte Carlo engine. In particular we have no grading scheme so we don't really have any way to compare the results with each other. But to implement that would likely mean significant changes to the internals, possibly a complete rework. Remember that this was mostly a research project up until now. Now that we're just about ready for production we may have to redesign some parts. Pity! :)

git-svn-id: svn+ssh://cat/usr/local/svnreps/skgb@631 31b05262-34a9-4923-903e-e8e313387d9b